inclusao de resoluções 1 a 20 lista decisão.

+++ secao_02_estrutura_de_decisao/ex_01_imprimir_maior_numero.py
@@ -16,10 +16,11 @@ Faça um Programa que peça dois números e imprima o maior deles.
 
 
 def maior_de_dois_numeros(x, y):
-    numero01 = float(input('Entre com o primeiro número: '))
-    numero02 = float(input('Entre com o segundo  número: '))
+    numero01 = int(input('Entre com o primeiro número: '))
+    numero02 = int(input('Entre com o segundo  número: '))
 
     if numero01 > numero02:
         print(numero01)
     else:
         print(numero02)
+
